requirements/base.txt
Django>=2.2,<2.3
django-filter>=2.1,<3
djangorestframework-jsonapi>=2.7,<3
